made some changes according to the pr
